requirements.txt
pandas
numpy
sklearn
django==3.2
django-dotenv==1.4.2
djangorestframework==3.13.1